Renamed files following pattern
