+++ src/pages/AlignmentAnalysis.js
@@ -7,7 +7,7 @@ import OlatcgSnackbar from "../components/OlatcgSnackbar";
 import useRequest from "../hooks/useRequest";
 import { API_ROUTES } from "../routes/Routes";
 import { getMessage } from "../services/MessageService";
-import Nodata from "../components/nodata";
+import OlatcgNodata from "../components/OlatcgOlatcgNodata";
 
 const AlignmentAnalysis = () => {
 
@@ -66,7 +66,7 @@ const AlignmentAnalysis = () => {
     
         return <> 
          
-            <Box sx={{px: 4, pb: 8}}>{info ? <Nodata />: 
+            <Box sx={{px: 4, pb: 8}}>{info ? <OlatcgNodata />: 
                 <Paper sx={{ width: '100%', overflow: 'hidden', bgcolor: 'primary.light' }}>
                     <TableContainer sx={{ maxHeight: '60vh' }}>
                         <Table stickyHeader aria-label="sticky table" >
